requirements.txt
blinker==1.7.0
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
promise==2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
singledispatch==3.7.0
six==1.16.0
SQLAlchemy==1.4.52
Werkzeug==3.0.2
zipp==3.18.1
